Move checkPermission method from QuestionController to QuestionService
