+++ src/main/java/com/codessquad/qna/service/QuestionService.java
@@ -1,10 +1,14 @@
 package com.codessquad.qna.service;
 
-import com.codessquad.qna.domain.Question;
-import com.codessquad.qna.exception.QuestionNotFoundException;
+import com.codessquad.qna.domain.*;
+import com.codessquad.qna.exception.*;
 import com.codessquad.qna.repository.QuestionRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import javax.servlet.http.HttpSession;
+
+import static com.codessquad.qna.controller.HttpSessionUtils.*;
 
 @Service
 public class QuestionService {
@@ -26,6 +30,22 @@ public class QuestionService {
 
     public Question findQuestion(Long id) {
         return questionRepository.findById(id).orElseThrow(QuestionNotFoundException::new);
+    }
+
+    public Question findQuestion(Long id, HttpSession session) {
+        Question question = questionRepository.findById(id).orElseThrow(QuestionNotFoundException::new);
+        checkPermission(session, question);
+        return question;
+    }
+
+    private void checkPermission(HttpSession session, Question question) {
+        if (!isLoginUser(session)) {
+            throw new FailedUserLoginException();
+        }
+        User loginUser = getSessionUser(session);
+        if (question.isNotSameAuthor(loginUser)) {
+            throw new IllegalUserAccessException();
+        }
     }
 }
